# specs/openapi.yaml
openapi: 3.1.0
info:
  title: FlowProof API
  version: 0.6.0
paths:
...
  /api/v1/chaos/mode:
    post:
      tags:
      - chaos
      summary: Set Chaos Mode
      operationId: set_chaos_mode_api_v1_chaos_mode_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaosModeIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Set Chaos Mode Api V1 Chaos Mode Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/principals:
    get:
      tags:
      - identity
      summary: List Principals
      operationId: list_principals_api_v1_identity_principals_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response List Principals Api V1 Identity Principals Get
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/humans:
    post:
      tags:
      - identity
      summary: Create Human
      operationId: create_human_api_v1_identity_humans_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HumanPrincipalIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Human Api V1 Identity Humans Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/service-accounts:
    post:
      tags:
      - identity
      summary: Create Service Account
      operationId: create_service_account_api_v1_identity_service_accounts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceAccountIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Service Account Api V1 Identity Service Accounts
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/principals/{principal_id}/role:
    post:
      tags:
      - identity
      summary: Change Role
      operationId: change_role_api_v1_identity_principals__principal_id__role_post
      parameters:
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          title: Principal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleChangeIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Change Role Api V1 Identity Principals  Principal
                  Id  Role Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/principals/{principal_id}/password:
    post:
      tags:
      - identity
      summary: Change Password
      operationId: change_password_api_v1_identity_principals__principal_id__password_post
      parameters:
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          title: Principal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/principals/{principal_id}/disable:
    post:
      tags:
      - identity
      summary: Disable Principal
      operationId: disable_principal_api_v1_identity_principals__principal_id__disable_post
      parameters:
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          title: Principal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/principals/{principal_id}/credentials:
    get:
      tags:
      - identity
      summary: List Credentials
      operationId: list_credentials_api_v1_identity_principals__principal_id__credentials_get
      parameters:
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          title: Principal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Credentials Api V1 Identity Principals  Principal
                  Id  Credentials Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
    post:
      tags:
      - identity
      summary: Issue Credential
      operationId: issue_credential_api_v1_identity_principals__principal_id__credentials_post
      parameters:
      - name: principal_id
        in: path
        required: true
        schema:
          type: string
          title: Principal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialIssueIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Issue Credential Api V1 Identity Principals  Principal
                  Id  Credentials Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/credentials/{credential_id}/rotate:
    post:
      tags:
      - identity
      summary: Rotate Credential
      operationId: rotate_credential_api_v1_identity_credentials__credential_id__rotate_post
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          title: Credential Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialRotateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Rotate Credential Api V1 Identity Credentials  Credential
                  Id  Rotate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/identity/credentials/{credential_id}/revoke:
    post:
      tags:
      - identity
      summary: Revoke Credential
      operationId: revoke_credential_api_v1_identity_credentials__credential_id__revoke_post
      parameters:
      - name: credential_id
        in: path
        required: true
        schema:
          type: string
          title: Credential Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
  /api/v1/security/audit:
    get:
      tags:
      - security
      summary: Security Audit
      operationId: security_audit_api_v1_security_audit_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: before
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Before
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Security Audit Api V1 Security Audit Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - BearerAuth: []
      - SessionCookie: []
...
components:
  schemas:
...
    ChaosModeIn:
      properties:
        mode:
          type: string
          pattern: ^(normal|false_200|timeout|delayed_write|amount_mismatch|partial_success)$
          title: Mode
        correlation_id:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Correlation Id
      type: object
      required:
      - mode
      title: ChaosModeIn
    CredentialIssueIn:
      properties:
        scopes:
          items:
            type: string
          type: array
          maxItems: 9
          minItems: 1
          uniqueItems: true
          title: Scopes
        expires_in_seconds:
          anyOf:
          - type: integer
            maximum: 2592000.0
            minimum: 1.0
          - type: 'null'
          title: Expires In Seconds
        label:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Label
      additionalProperties: false
      type: object
      required:
      - scopes
      title: CredentialIssueIn
    CredentialRotateIn:
      properties:
        expires_in_seconds:
          anyOf:
          - type: integer
            maximum: 2592000.0
            minimum: 1.0
          - type: 'null'
          title: Expires In Seconds
      additionalProperties: false
      type: object
      title: CredentialRotateIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HumanPrincipalIn:
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        password:
          type: string
          maxLength: 1024
          minLength: 12
          title: Password
        role:
          type: string
          pattern: ^(viewer|operator|admin)$
          title: Role
      additionalProperties: false
      type: object
      required:
      - name
      - password
      - role
      title: HumanPrincipalIn
...
    PasswordChangeIn:
      properties:
        password:
          type: string
          maxLength: 1024
          minLength: 12
          title: Password
      additionalProperties: false
      type: object
      required:
      - password
      title: PasswordChangeIn
...
    RoleChangeIn:
      properties:
        role:
          type: string
          pattern: ^(viewer|operator|admin)$
          title: Role
      additionalProperties: false
      type: object
      required:
      - role
      title: RoleChangeIn
    ServiceAccountIn:
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 1
          title: Name
        scopes:
          items:
            type: string
          type: array
          maxItems: 9
          minItems: 1
          uniqueItems: true
          title: Scopes
      additionalProperties: false
      type: object
      required:
      - name
      - scopes
      title: ServiceAccountIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API credential issued to a human or service principal.
    SessionCookie:
      type: apiKey
      in: cookie
      name: flowproof_session
      description: Human browser session; state-changing requests also require X-CSRF-Token.